feat(autolist) : add initial commit for autolist
Add autolist for <CR>.
if show_marker is true, it makes new marker to next line.
if show_marker is false, it doesn't make new marker but the place which
would marker is will be substitute by white space to next line.
it is same behavior with <S-CR> in other GUI program like word, ppt.

But <S-CR> cannot be mapped because it is recognized like <CR> in most
of terminal.
So <M-CR> is used instead of it. to use case that show_marker is false.