feat(links) : `follow_link()` can open other file which is not image
1. Problem

If the link is not image file like .pdf, .word etc...
`follow_link()` cannot open the file. because it connects to
vim.lsp.definition()

2. Solution

Check the link means internal file first.
If not, check the link is image file. if yes, opens with `image_opener`.
If not, opens with `file_opener`

so rename `file_opener` to `md_opener` which means that vim command
before file open.
`file_opener` means opener which opens external file.